// src/data/api/datatyper/personalia.tsx
export interface Vegadresse {
    matrikkelId: OrNothing<number>;
    postnummer: StringOrNothing;
    husnummer: StringOrNothing;
    husbokstav: StringOrNothing;
    kommunenummer: StringOrNothing;
    kommune: StringOrNothing;
    adressenavn: StringOrNothing;
    tilleggsnavn: StringOrNothing;
    poststed: StringOrNothing;
}

export interface Matrikkeladresse {
    matrikkelId: OrNothing<number>;
    bruksenhetsnummer: StringOrNothing;
    tilleggsnavn: StringOrNothing;
    postnummer: StringOrNothing;
    kommunenummer: StringOrNothing;
    kommune: StringOrNothing;
    poststed: StringOrNothing;
}

export interface Utenlandskadresse {
    adressenavnNummer: StringOrNothing;
    bygningEtasjeLeilighet: StringOrNothing;
    postboksNummerNavn: StringOrNothing;
    postkode: StringOrNothing;
    bySted: StringOrNothing;
    regionDistriktOmraade: StringOrNothing;
    landkode: StringOrNothing;
}

export interface Ukjentbosted {
    bostedskommune: StringOrNothing;
    kommune: StringOrNothing;
}

export interface Postboksadresse {
    postbokseier: StringOrNothing;
    postboks: StringOrNothing;
    postnummer: StringOrNothing;
    poststed: StringOrNothing;
}

export interface PostadresseIFrittFormat {
    adresselinje1: StringOrNothing;
    adresselinje2: StringOrNothing;
    adresselinje3: StringOrNothing;
    postnummer: StringOrNothing;
    poststed: StringOrNothing;
}

export interface UtenlandskadresseIFrittFormat {
    adresselinje1: StringOrNothing;
    adresselinje2: StringOrNothing;
    adresselinje3: StringOrNothing;
    byEllerStedsnavn: StringOrNothing;
    postkode: StringOrNothing;
    landkode: StringOrNothing;
}
// ...
export interface Kontaktadresse {
    type: String;
    coAdressenavn: StringOrNothing;
    vegadresse: OrNothing<Vegadresse>;
    postboksadresse: OrNothing<Postboksadresse>;
    utenlandskAdresse: OrNothing<Utenlandskadresse>;
    postadresseIFrittFormat: OrNothing<PostadresseIFrittFormat>;
    utenlandskAdresseIFrittFormat: OrNothing<UtenlandskadresseIFrittFormat>;
}